+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Support `composer/installers` v1 and v2
 - Link to settings page when foreign drop-in was found
 - Added `wp_cache_flush_runtime()` function
-- Added `wp_cache_delete_multiple()` function
+- Added `wp_cache_*delete*_multiple()` functions
 
 ## 2.0.23